setup.py auto commit

+++ src/MMCQsc/scp/main.py
@@ -95,7 +95,7 @@ def openhtml(myip,PORT):
         logger.info(f'\n\n\n\t\t即将默认浏览器打开：\n\n\t\t http://{myip}:{PORT}\n\n')
         os.system(f'start http://{myip}:{PORT}')
 
-def mainFunc():
+def mainFunc(mode=False):
     from MMCQsc.scp.lib import dpkg
     pgd = dpkg.Pgd(BASE_DIR,DPKG_DIR)
     new = dpkg.check_update(f"{__version__}","https://pypi.org/pypi/color-theme-analyse/json","https://mirrors.tencent.com/pypi/simple/color-theme-analyse/")
@@ -141,15 +141,29 @@ def mainFunc():
                 from MMCQsc.scp.scripts import executer
                 result = executer.domain(img)
                 if buf[2] == 1 and result == 6:
-                    with futures.ProcessPoolExecutor(max_workers=None) as prolist:
-                        PORT = randint(5800,5858)
-                        if os.name != 'posix':
-                            myip = get_host_ip()
+                    if mode:
+                        if os.name == 'posix':
+                            _browser = os.path.abspath(os.path.join(SRC_DIR, 'browser','linux'))
                         else:
-                            myip = 'localhost'
-                        prolist.submit(createServer,myip,PORT)
+                            _browser = os.path.abspath(os.path.join(SRC_DIR, 'browser','windows'))
+                        # response = os.system(f'{_browser} \"{SRC_DIR}\"')
+                        _index = SRC_DIR.replace('\\','/')
+                        args = ['TaskBar',f'{_index}']
+                        print(args)
+                        input()
+                        Popen(args, bufsize=0, close_fds=False, shell=True, cwd=_browser, startupinfo=None, creationflags=0)
                         time.sleep(2)
-                        prolist.submit(openhtml,myip,PORT)  # 多进程才能打开
+                        exit(9)
+                    else:
+                        with futures.ProcessPoolExecutor(max_workers=None) as prolist:
+                            PORT = randint(5800,5858)
+                            if os.name != 'posix':
+                                myip = get_host_ip()
+                            else:
+                                myip = 'localhost'
+                            prolist.submit(createServer,myip,PORT)
+                            time.sleep(2)
+                            prolist.submit(openhtml,myip,PORT)  # 多进程才能打开
             else:
                 logger.error("无输入或无效输入")
                 shm.close()
@@ -161,11 +175,12 @@ def mainFunc():
                 logger.debug(e)
         finally:
             try:
-                os.remove(os.path.join(SRC_DIR, 'index.js'))
-                os.remove(os.path.join(SRC_DIR, 'index.css'))
-                shutil.rmtree(os.path.join(SRC_DIR, 'finish'))
-                shutil.rmtree(os.path.join(SRC_DIR, 'compress'))
-                logger.info('成功删除不重要的自动生成文件')
+                if mode == False:
+                    os.remove(os.path.join(SRC_DIR, 'index.js'))
+                    os.remove(os.path.join(SRC_DIR, 'index.css'))
+                    shutil.rmtree(os.path.join(SRC_DIR, 'finish'))
+                    shutil.rmtree(os.path.join(SRC_DIR, 'compress'))
+                    logger.info('成功删除不重要的自动生成文件')
                 logger.warning("\n\n\t\t[ tip ] : 如果当前窗口未正确返回 Shell 环境，使用 CTRL + PAUSE_BREAK 强制结束所有任务并退出 Python\n\n")
             except Exception as e:
                 if isinstance(e,FileNotFoundError):
